fix(immutable-arraybuffer): unify shim to work on more platforms
