+++ packages/immutable-arraybuffer/package.json
@@ -25,7 +25,6 @@
   "exports": {
     ".": "./index.js",
     "./shim.js": "./shim.js",
-    "./shim-hermes.js": "./shim-hermes.js",
     "./package.json": "./package.json"
   },
   "scripts": {
