+++ packages/immutable-arraybuffer/test/immutable-arraybuffer-shim-transfer.test.js
@@ -1,20 +1,18 @@
 // @ts-nocheck
 import test from 'ava';
-import {
-  isBufferImmutable,
-  sliceBufferToImmutable,
-} from '../src/limited-pony-for-hermes.js';
+import '../src/immutable-arraybuffer-shim.js';
 
 const { isFrozen, getPrototypeOf } = Object;
 
-test('Immutable ArrayBuffer ponyfill installed and not hardened', t => {
-  const iab = sliceBufferToImmutable(new ArrayBuffer(0));
+test('Immutable ArrayBuffer shim installed but not hardened', t => {
+  const ab1 = new ArrayBuffer(0);
+  const iab = ab1.transferToImmutable();
   const iabProto = getPrototypeOf(iab);
   t.false(isFrozen(iabProto));
   t.false(isFrozen(iabProto.slice));
 });
 
-test('Immutable ArrayBuffer ponyfill ops', t => {
+test('Immutable ArrayBuffer shim ops', t => {
   // Absent on Node <= 18
   const canResize = 'maxByteLength' in ArrayBuffer.prototype;
 
@@ -22,12 +20,12 @@ test('Immutable ArrayBuffer ponyfill ops', t => {
   const ta1 = new Uint8Array(ab1);
   ta1[0] = 3;
   ta1[1] = 4;
-  const iab = sliceBufferToImmutable(ab1);
+  const iab = ab1.transferToImmutable();
   t.true(iab instanceof ArrayBuffer);
   ta1[1] = 5;
   const ab2 = iab.slice(0);
   const ta2 = new Uint8Array(ab2);
-  t.is(ta1[1], 5);
+  t.is(ta1[1], undefined);
   t.is(ta2[1], 4);
   ta2[1] = 6;
 
@@ -35,22 +33,22 @@ test('Immutable ArrayBuffer ponyfill ops', t => {
   t.true(ab3 instanceof ArrayBuffer);
 
   const ta3 = new Uint8Array(ab3);
-  t.is(ta1[1], 5);
+  t.is(ta1[1], undefined);
   t.is(ta2[1], 6);
   t.is(ta3[1], 4);
 
-  t.is(ab1.byteLength, 2);
+  t.is(ab1.byteLength, 0);
   t.is(iab.byteLength, 2);
   t.is(ab2.byteLength, 2);
 
   t.is(iab.maxByteLength, 2);
   if (canResize) {
-    t.is(ab1.maxByteLength, 7);
+    t.is(ab1.maxByteLength, 0);
     t.is(ab2.maxByteLength, 2);
   }
 
   if ('detached' in ab1) {
-    t.false(ab1.detached);
+    t.true(ab1.detached);
     t.false(ab2.detached);
     t.false(ab3.detached);
   }
@@ -73,13 +71,13 @@ test('Standard DataView behavior baseline', t => {
 // This could have been written as a test.failing as compared to
 // the immutable ArrayBuffer we'll propose. However, I'd rather test what
 // the shim purposely does instead.
-test('DataView on Immutable ArrayBuffer ponyfill limitations', t => {
+test('DataView on Immutable ArrayBuffer shim limitations', t => {
   const ab1 = new ArrayBuffer(2);
   const ta1 = new Uint8Array(ab1);
   ta1[0] = 3;
   ta1[1] = 4;
 
-  const iab = sliceBufferToImmutable(ab1);
+  const iab = ab1.transferToImmutable();
   t.throws(() => new DataView(iab), {
     instanceOf: TypeError,
   });
@@ -105,7 +103,7 @@ test('Standard TypedArray behavior baseline', t => {
 // This could have been written as a test.failing as compared to
 // the immutable ArrayBuffer we'll propose. However, I'd rather test what
 // the shim purposely does instead.
-test('TypedArray on Immutable ArrayBuffer ponyfill limitations', t => {
+test('TypedArray on Immutable ArrayBuffer shim limitations', t => {
   const ab1 = new ArrayBuffer(2);
   const dv1 = new DataView(ab1);
   t.is(dv1.buffer, ab1);
@@ -115,7 +113,7 @@ test('TypedArray on Immutable ArrayBuffer ponyfill limitations', t => {
   ta1[1] = 4;
   t.is(ta1.byteLength, 2);
 
-  const iab = sliceBufferToImmutable(ab1);
+  const iab = ab1.transferToImmutable();
   // Unfortunately, unlike the immutable ArrayBuffer to be proposed,
   // calling a TypedArray constructor with the shim implementation of
   // an immutable ArrayBuffer as argument treats it as an unrecognized object,
@@ -132,7 +130,7 @@ const testTransfer = t => {
   t.deepEqual([...ta12], [3, 4, 5]);
 
   const ab2 = ab12.transfer(5);
-  t.false(isBufferImmutable(ab2));
+  t.false(ab2.immutable);
   t.is(ab2.byteLength, 5);
   t.is(ab12.byteLength, 0);
   const ta2 = new Uint8Array(ab2);
@@ -142,7 +140,7 @@ const testTransfer = t => {
   const ab13 = ta13.buffer;
 
   const ab3 = ab13.transfer(2);
-  t.false(isBufferImmutable(ab3));
+  t.false(ab3.immutable);
   t.is(ab3.byteLength, 2);
   t.is(ab13.byteLength, 0);
   const ta3 = new Uint8Array(ab3);
@@ -155,48 +153,48 @@ const testTransfer = t => {
   maybeTest('Standard buf.transfer(newLength) behavior baseline', testTransfer);
 }
 
-test('Analogous sliceBufferToImmutable(buf, 0, newLength) ponyfill', t => {
+test('Analogous buf.transferToImmutable(newLength) shim', t => {
   const ta12 = new Uint8Array([3, 4, 5]);
   const ab12 = ta12.buffer;
   t.is(ab12.byteLength, 3);
   t.deepEqual([...ta12], [3, 4, 5]);
 
-  const ab2 = sliceBufferToImmutable(ab12, 0, 5);
-  t.true(isBufferImmutable(ab2));
-  t.is(ab2.byteLength, 3);
-  t.is(ab12.byteLength, 3);
+  const ab2 = ab12.transferToImmutable(5);
+  t.true(ab2.immutable);
+  t.is(ab2.byteLength, 5);
+  t.is(ab12.byteLength, 0);
   // slice needed due to ponyfill limitations.
   const ta2 = new Uint8Array(ab2.slice());
-  t.deepEqual([...ta2], [3, 4, 5]);
+  t.deepEqual([...ta2], [3, 4, 5, 0, 0]);
 
   const ta13 = new Uint8Array([3, 4, 5]);
   const ab13 = ta13.buffer;
 
-  const ab3 = sliceBufferToImmutable(ab13, 0, 2);
-  t.true(isBufferImmutable(ab3));
+  const ab3 = ab13.transferToImmutable(2);
+  t.true(ab3.immutable);
   t.is(ab3.byteLength, 2);
-  t.is(ab13.byteLength, 3);
+  t.is(ab13.byteLength, 0);
   // slice needed due to ponyfill limitations.
   const ta3 = new Uint8Array(ab3.slice());
   t.deepEqual([...ta3], [3, 4]);
 });
 
-test('sliceBufferToImmutable ponyfill', t => {
+test('sliceToImmutable shim', t => {
   const ta12 = new Uint8Array([3, 4, 5]);
   const ab12 = ta12.buffer;
   t.is(ab12.byteLength, 3);
   t.deepEqual([...ta12], [3, 4, 5]);
 
-  const ab2 = sliceBufferToImmutable(ab12, 1, 5);
-  t.true(isBufferImmutable(ab2));
+  const ab2 = ab12.sliceToImmutable(1, 5);
+  t.true(ab2.immutable);
   t.is(ab2.byteLength, 2);
   t.is(ab12.byteLength, 3);
   // slice needed due to ponyfill limitations.
   const ta2 = new Uint8Array(ab2.slice());
   t.deepEqual([...ta2], [4, 5]);
 
-  const ab3 = sliceBufferToImmutable(ab2, 1, 2);
-  t.true(isBufferImmutable(ab3));
+  const ab3 = ab2.sliceToImmutable(1, 2);
+  t.true(ab3.immutable);
   t.is(ab3.byteLength, 1);
   t.is(ab2.byteLength, 2);
   // slice needed due to ponyfill limitations.
